stable-backend/worker/requirements.txt
accelerate==0.14.0
async-timeout==4.0.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
diffusers==0.12.1
filelock==3.8.0
Flask==2.2.2
ftfy==6.1.1
huggingface-hub==0.11.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.4
packaging==21.3
Pillow==9.3.0
psutil==5.9.4
pyparsing==3.0.9
python-dotenv==1.0.0
PyYAML==6.0
redis==4.3.4
regex==2022.10.31
requests==2.28.1
rq==1.11.1
scipy==1.9.3
tokenizers==0.13.2
tqdm==4.64.1
transformers==4.24.0
typing-extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
